State Wrappers
Added complete documentation
