+++ src/tui/mod.rs
@@ -29,7 +29,5 @@ pub mod modes {
     pub mod backend_interface;
     pub mod block_data;
     pub mod keybinding_manager;
-    pub mod table_state_wrapper;
-    pub mod list_state_wrapper;
-    // pub mod tui_widget;
+    pub mod state_wrappers;
 }
